start now-node200-loopback-workshop
We are going to build out an API that will allow CRUD operations for
Teams and Players, and query for players that belong to teams. This
project is concerned with the back end, and doesn't need to have a
front facing web app. The purpose of this workshop is to demonstrate
how using a framework can make building and maintaining APIs easy.